Novos exercícios feitos e ajustado erros em outros exercícios

+++ src/com/example/exercicios/Exercicio5.kt
@@ -4,7 +4,7 @@ fun main(){
     val n1: String = readLine().toString()
     val n2: String = readLine().toString()
 
-    val media: Double = (n1.toDouble() + n2.toDouble())/2
+    val media: Float = (n1.toFloat() + n2.toFloat())/2
 
     println("MEDIA = $media")
 }
